requirements.txt
streamlit
python-dotenv
google-generativeai
openai
pandas
plotly
yfinance
streamlit-pdf-viewer